fix(importer): handle bad-CIDR rows and 4-byte ASNs; ship schema migrations in chart
The fresh-postgres recovery on the irrexplorer cluster surfaced three latent bugs:

1. BGP importer crashed on 4-byte ASNs because the bgp.asn column was int4
   and BGPEntry.ASN was Go int. Both are now bigint / int64.

2. RIR stats importer aborted the entire batch on a single misaligned CIDR
   row from upstream delegation files (pgx CopyFrom is all-or-nothing).
   Prefixes are now validated and masked via netip.ParsePrefix; unparseable
   rows are silently skipped.

3. The repo shipped no schema-bootstrap migration — the original schema was
   created implicitly by the legacy Python codebase. Added 000_init.sql
   alongside the existing 001_drop_alerting_tables.sql under
   charts/irrexplorer/migrations/, and rewrote the helm migrate-job to
   apply every .sql file in lexical order via a ConfigMap. All migrations
   are idempotent.

Removed the orphaned go-backend/db/migrations/ directory (single source of
truth now lives inside the chart).

+++ go-backend/internal/importer/bgp.go
@@ -16,9 +16,12 @@ const bgpToolsURL = "https://bgp.tools/table.jsonl"
 
 // BGPEntry is one line from bgp.tools/table.jsonl.
 // bgp.tools uses uppercase keys: {"CIDR":"...","ASN":...,"Hits":...}
+// ASN is int64 because 4-byte ASNs (e.g. 4200001401) exceed the 32-bit signed
+// range and pgx COPY rejects them when the destination column is int4. The
+// matching DB column is bigint.
 type BGPEntry struct {
 	Prefix string `json:"CIDR"`
-	ASN    int    `json:"ASN"`
+	ASN    int64  `json:"ASN"`
 }
 
 // ParseBGPLine parses a single JSONL line. Exported for testing.
